Store data like keywords in local database
This patch adds support for the database drivers which can e.g. store
already known keywords in the database.
For now only implemented driver is "local_file" driver which simply
stores all known keywords in the local file, one tag per line.
In the future this driver may be replaced by e.g. sqlite3 or something
like that if there will be need to store more data in local database.

Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>